+++ frontend/src/components/components/Navbars/AdminNavbarLinks.js
@@ -1,3 +1,4 @@
+import React, {useRef} from "react";
 // Chakra Icons
 import {BellIcon, SearchIcon} from "@chakra-ui/icons";
 // Chakra Imports
@@ -25,7 +26,7 @@ import {ProfileIcon, SettingsIcon} from "../Icons/Icons";
 import {ItemContent} from "../Menu/ItemContent";
 import SidebarResponsive from "../Sidebar/SidebarResponsive";
 import PropTypes from "prop-types";
-import React from "react";
+
 import {NavLink, useHistory} from "react-router-dom";
 import {useDispatch} from "react-redux";
 import routes from "../../../routes.js";
@@ -43,11 +44,17 @@ export default function HeaderLinks(props) {
     let navbarIcon = useColorModeValue("gray.500", "gray.200");
     let searchIcon = useColorModeValue("gray.700", "gray.200");
 
+    const handleExit = () => {
+        dispatch(logout());
+        history.push('/');
+        window.location.reload();
+    }
+
     if (secondary) {
         navbarIcon = "white";
         mainText = "white";
     }
-    const settingsRef = React.useRef();
+    const settingsRef = useRef();
     return (
         <Flex
             pe={{sm: "0px", md: "16px"}}
@@ -100,7 +107,7 @@ export default function HeaderLinks(props) {
             <Button
                 ms="0px"
                 px="0px"
-                onClick={() => {localStorage.removeItem('userInfo'); history.push('/')}}
+                onClick={handleExit}
                 me={{sm: "2px", md: "16px"}}
                 color={navbarIcon}
                 variant="transparent-with-icon"