+++ frontend/src/screens/HomeScreen.js
@@ -1,7 +1,9 @@
 import React from 'react';
 
 import {Container} from "react-bootstrap";
-import {useSelector} from "react-redux";
+import {useDispatch} from "react-redux";
+
+import {useHistory} from "react-router";
 
 import "../css/main.css";
 import "../css/buttons.css";
@@ -23,10 +25,7 @@ import ListOfRec from "../components/ListOfRec";
 
 import dataForAccordion from "../helpers/dataForAccordion";
 
-
 const HomeScreen = () => {
-
-    const user = localStorage.getItem('userInfo');
 
     return (
         <div id="body">
@@ -34,7 +33,7 @@ const HomeScreen = () => {
             <Header/>
             <Container>
                 <div id="first-block">
-                    <div className="first-block-text" onClick={() => console.log(user)}>
+                    <div className="first-block-text">
                         <h2> Негативные отзывы теперь в прошлом!</h2>
                         <h1><b> <span className="gradient-text">RecTop</span> - инновационный сервис</b> <span
                             className="cursive-text special">взаимодействия с клиентами.</span></h1>